Rename commons:test to commonds:testing
With this old name it seems like IntelliJ something cannot index
the classes inside this project.
