+++ build.gradle
@@ -66,7 +66,7 @@ configure(subprojects.findAll { it.name.endsWith('-impl') } ) {
         runtimeOnly 'com.h2database:h2'
 
         // Test
-        testImplementation project(':commons:test')
+        testImplementation project(':commons:testing')
         testImplementation 'org.springframework.boot:spring-boot-starter-test'
         testCompile 'org.springframework.restdocs:spring-restdocs-mockmvc'
     }
